refactor: CarsDto 생성
