+++ src/main/java/racingcar/Application.java
@@ -5,6 +5,7 @@ import java.util.stream.Collectors;
 
 import racingcar.domain.Cars;
 import racingcar.dto.CarDto;
+import racingcar.dto.CarsDto;
 import racingcar.strategy.RandomMovableStrategy;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
@@ -12,7 +13,8 @@ import racingcar.view.OutputView;
 public class Application {
     public static void main(String[] args) {
         InputView inputView = new InputView();
-        Cars cars = Cars.from(inputView.inputCarNames());
+        CarsDto carsDto = inputView.inputCarNames();
+        Cars cars = carsDto.toEntity();
         int tryCount = inputView.inputTryCount();
         startGame(cars, tryCount);
     }
