--- src/main/java/racingcar/dto/CarNameDto.java
@@ -1,30 +1,0 @@
-package racingcar.dto;
-
-import java.util.Objects;
-
-public class CarNameDto {
-    private final String name;
-
-    public CarNameDto(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (o == null || getClass() != o.getClass())
-            return false;
-        CarNameDto other = (CarNameDto)o;
-        return Objects.equals(getName(), other.getName());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(getName());
-    }
-}
